Update edited document description
